requirements.txt
ffmpeg-python
yt-dlp
tafrigh
